Split @Composable UI into several files
